refactor: colocate home-page search components under app/(home)
